+++ networkReliabilityCommon/obs/withResampling.cpp
@@ -20,7 +20,7 @@ namespace networkReliability
 			::networkReliability::NetworkReliabilityObs::constructConditional(context, randomSource, state.get(), false);
 
 			const ::TruncatedBinomialDistribution::TruncatedBinomialDistribution& originalDist = context.getOpDistribution(0, nEdges, nEdges);
-			mpfr_class conditioningProb = originalDist.getCumulativeProbability(nEdges - minCutEdges);
+			mpfr_class conditioningProb = originalDist.getCumulativeProbability((int)nEdges - (int)minCutEdges);
 			return withResampling(context, state, (int)minCutEdges, conditioningProb);
 		}
 		withResampling::withResampling(Context const& context, boost::shared_array<EdgeState> state, ::networkReliability::obs::withResamplingConstructorType& other)
